Fix capitalization of hostname in prompt on WSL
WSL uses windows hostnames which begin with a capital no matter how they
are entered. zsh prompt expansion does not provide a way to change
punctuation, so used the 'hostname' command instead. This may not be
available on all machines. On Arch Linux this command is part of the
package 'core/inteutils'.