Add better mechanism to autofetch changes
As per commit 1d2b395b54f8f289a0a248f864928dad45438dd8, this is a better
method of trying to fetch changes to dotfiles automatically in the
background. Currently this will trigger whenever zshrc is sourced, which
may not be ideal. If this seems to be overly aggressive, consideration
will be given to either trying to find a resolution or just scrapping
the whole idea.