Change virtual movement to affect normal mode only
When config mappings were changes so line movement was based on
virutal-lines (displayed lines instead of physical lines; big difference
when text is wrapped), they were applied to all modes. This is not
always optimal and more often than not outside of normal mode the
default behaviour is more desirable. This changes those mappings to
affect normal mode only.