Make Steam play nicely and other small changes
The external rules file for "spotify" was already looking at the edge
case where no WM_NAME exists or where it is set too late for
conventional rules to work. As Steam (particularly the update window on
launch) also doesn't set properly, it made sense to make it a more
general external rule and to adequately account for steam as well. While
I was at it, I added Steam and the emacs daemon to the list of
applications that start automatically boot.