Improve wlan support in polybar
For some reason, wlan0 doesn't always load at boot on my laptop and I
currently don't have the time to troubleshoot it. When the appropriate
kernel modules are reloaded, everything works fine but the device is
renamed wlan1. All connected wlan's will now be displayed. Also added a
simple script to remove and reload the modules as needed.