Add alias for cp to be physical and not symbolic
If changing into a physical directory, the '-P' option has no effect and
generates no errors. When following a symbolic link to change
directories, the '-P' option will follow the physical directory pointed
to by the symbolic link which is preferred.

For example, if '$HOME/test' is a symbolic link pointing to
'/usr/share', the value of '$PWD' will be '$HOME/test' after 'cd -P
$HOME/test'. Generally it would be preferred for '$PWD' to be
'/usr/share'. To override this, given that cd is a shell built-in, use
'\cd'.