Add ability to cancel alarm
Previously, once an alarm had started there was no way to end execution
easily except by killing all instances of mpg123. While muscle memory
could be developed to do this simply, if the function was ever changed
to use something else, the user would need to relearn that muscle
memory. It also requires knowing that which executable is being used. By
using a link to the executable with a unique and predictable name, the
function can be changed at will and the actual media player used for the
alarm is irrelevant.